+++ mithril-common/src/chain_reader/pallas_chain_reader.rs
@@ -67,18 +67,16 @@ impl PallasChainReader {
         next: NextResponse<BlockContent>,
     ) -> StdResult<Option<ChainBlockNextAction>> {
         match next {
-            NextResponse::RollForward(raw_block, forward_tip) => {
+            NextResponse::RollForward(raw_block, _forward_tip) => {
                 let multi_era_block = MultiEraBlock::decode(&raw_block)
                     .with_context(|| "PallasChainReader failed to decode raw block")?;
                 let parsed_block = ScannedBlock::convert(multi_era_block, 0);
-                Ok(Some(ChainBlockNextAction::RollForward {
-                    next_point: forward_tip.into(),
-                    parsed_block,
-                }))
+                Ok(Some(ChainBlockNextAction::RollForward { parsed_block }))
             }
             NextResponse::RollBackward(rollback_point, _) => {
+                let chain_point = ChainPoint::from(rollback_point);
                 Ok(Some(ChainBlockNextAction::RollBackward {
-                    rollback_point: rollback_point.into(),
+                    slot_number: chain_point.slot_number,
                 }))
             }
             NextResponse::Await => Ok(None),
@@ -162,11 +160,6 @@ mod tests {
     /// Returns a fake chain point for testing purposes.
     fn get_fake_chain_point_backwards() -> ChainPoint {
         ChainPoint::from(get_fake_specific_point())
-    }
-
-    /// Returns a fake chain point for testing purposes.
-    fn get_fake_chain_point_forwards() -> ChainPoint {
-        Tip(get_fake_specific_point(), get_fake_block_number()).into()
     }
 
     /// Creates a new work directory in the system's temporary folder.
@@ -272,8 +265,8 @@ mod tests {
         let (_, client_res) = tokio::join!(server, client);
         let chain_block = client_res.expect("Client failed to get next chain block");
         match chain_block {
-            ChainBlockNextAction::RollBackward { rollback_point } => {
-                assert_eq!(rollback_point, get_fake_chain_point_backwards());
+            ChainBlockNextAction::RollBackward { slot_number } => {
+                assert_eq!(slot_number, get_fake_chain_point_backwards().slot_number);
             }
             _ => panic!("Unexpected chain block action"),
         }
@@ -304,11 +297,7 @@ mod tests {
         let (_, client_res) = tokio::join!(server, client);
         let chain_block = client_res.expect("Client failed to get next chain block");
         match chain_block {
-            ChainBlockNextAction::RollForward {
-                next_point,
-                parsed_block,
-            } => {
-                assert_eq!(next_point, get_fake_chain_point_forwards());
+            ChainBlockNextAction::RollForward { parsed_block } => {
                 assert_eq!(parsed_block, get_fake_scanned_block());
             }
             _ => panic!("Unexpected chain block action"),
@@ -344,11 +333,7 @@ mod tests {
         let (_, client_res) = tokio::join!(server, client);
         let chain_block = client_res.expect("Client failed to get next chain block");
         match chain_block {
-            ChainBlockNextAction::RollForward {
-                next_point,
-                parsed_block,
-            } => {
-                assert_eq!(next_point, get_fake_chain_point_forwards());
+            ChainBlockNextAction::RollForward { parsed_block } => {
                 assert_eq!(parsed_block, get_fake_scanned_block());
             }
             _ => panic!("Unexpected chain block action"),