+++ mithril-signer/src/cardano_transactions_importer.rs
@@ -8,7 +8,9 @@ use slog::{debug, Logger};
 
 use mithril_common::cardano_block_scanner::{BlockScanner, ChainScannedBlocks};
 use mithril_common::crypto_helper::{MKTree, MKTreeNode};
-use mithril_common::entities::{BlockNumber, BlockRange, CardanoTransaction, ChainPoint};
+use mithril_common::entities::{
+    BlockNumber, BlockRange, CardanoTransaction, ChainPoint, SlotNumber,
+};
 use mithril_common::signable_builder::TransactionsImporter;
 use mithril_common::StdResult;
 
@@ -41,11 +43,11 @@ pub trait TransactionStore: Send + Sync {
 
     /// Remove transactions and block range roots that are in a rolled-back fork
     ///
-    /// * Remove transactions with block number strictly greater than the given block number
-    /// * Remove block range roots that have lower bound range strictly above the given block number
+    /// * Remove transactions with slot number strictly greater than the given slot number
+    /// * Remove block range roots that have lower bound range strictly above the given slot number
     async fn remove_rolled_back_transactions_and_block_range(
         &self,
-        block_number: BlockNumber,
+        slot_number: SlotNumber,
     ) -> StdResult<()>;
 }
 
@@ -115,9 +117,9 @@ impl CardanoTransactionsImporter {
                         .store_transactions(parsed_transactions)
                         .await?;
                 }
-                ChainScannedBlocks::RollBackward(chain_point) => {
+                ChainScannedBlocks::RollBackward(slot_number) => {
                     self.transaction_store
-                        .remove_rolled_back_transactions_and_block_range(chain_point.block_number)
+                        .remove_rolled_back_transactions_and_block_range(slot_number)
                         .await?;
                 }
             }
@@ -701,7 +703,7 @@ mod tests {
             .await
             .unwrap();
 
-        let chain_point = ChainPoint::new(1, 130, "block_hash-131");
+        let chain_point = ChainPoint::new(5, 130, "block_hash-131");
         let scanner = DumbBlockScanner::new().backward(chain_point);
 
         let importer =
@@ -748,6 +750,19 @@ mod tests {
                 .iter()
                 .map(|b| (b.clone(), MKTreeNode::from_hex("AAAA").unwrap()))
                 .collect(),
+            )
+            .await
+            .unwrap();
+        repository
+            .store_transactions(
+                ScannedBlock::new(
+                    "block_hash-131",
+                    BlockRange::from_block_number(BlockRange::LENGTH * 3).start,
+                    1,
+                    0,
+                    vec!["tx_hash-1", "tx_hash-2", "tx_hash-3"],
+                )
+                .into_transactions(),
             )
             .await
             .unwrap();